+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Road map
 
 - [x] Removed check for TypeScript capability on CPS
+- [x] Disable command when solution is building or debugging
+- [ ] Move strings to resource file
 - [ ] Minification of the generated .js
 
 Features that have a checkmark are complete and available for